requirements.txt
aiogram==3.22.0
yt-dlp==2025.8.1
aiohttp==3.8.5
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.3
pydub
